+++ src/store/modules/User.js
@@ -53,6 +53,90 @@ export default {
         setPreferenceSoundEffectsVolume(state, volume) {
             state.preferences.audio.soundEffects.volume = volume;
         },
+        setPreferencesGameOptions(state, gameOptions) {
+            state.preferences.game.options = gameOptions;
+        },
+        setPreferenceGameOptionRepartitionIsHidden(state, isHidden) {
+            state.preferences.game.options.repartition.isHidden = isHidden;
+        },
+        setPreferenceGameOptionAreRolesRevealedOnDeath(state, areRolesRevealedOnDeath) {
+            state.preferences.game.options.roles.areRevealedOnDeath = areRolesRevealedOnDeath;
+        },
+        setPreferenceGameOptionIsSheriffEnabled(state, isSheriffEnabled) {
+            state.preferences.game.options.roles.sheriff.isEnabled = isSheriffEnabled;
+        },
+        setPreferenceGameOptionSheriffElectedAtTurn(state, sheriffElectedAtTurn) {
+            state.preferences.game.options.roles.sheriff.electedAt.turn = sheriffElectedAtTurn;
+        },
+        setPreferenceGameOptionSheriffElectedAtPhase(state, sheriffElectedAtPhase) {
+            state.preferences.game.options.roles.sheriff.electedAt.phase = sheriffElectedAtPhase;
+        },
+        setPreferenceGameOptionIsSheriffVoteDoubled(state, isSheriffVoteDoubled) {
+            state.preferences.game.options.roles.sheriff.hasDoubledVote = isSheriffVoteDoubled;
+        },
+        setPreferenceGameOptionIsBigBadWolfPowerlessIfWerewolfDies(state, isBigBadWolfPowerlessIfWerewolfDies) {
+            state.preferences.game.options.roles.bigBadWolf.isPowerlessIfWerewolfDies = isBigBadWolfPowerlessIfWerewolfDies;
+        },
+        setPreferenceGameOptionWhiteWerewolfWakingUpInterval(state, whiteWerewolfWakingUpInterval) {
+            state.preferences.game.options.roles.whiteWerewolf.wakingUpInterval = whiteWerewolfWakingUpInterval;
+        },
+        setPreferenceGameOptionIsSeerTalkative(state, isSeerTalkative) {
+            state.preferences.game.options.roles.seer.isTalkative = isSeerTalkative;
+        },
+        setPreferenceGameOptionCanSeerSeeRoles(state, canSeerSeeRoles) {
+            state.preferences.game.options.roles.seer.canSeeRoles = canSeerSeeRoles;
+        },
+        setPreferenceGameOptionIsLittleGirlProtectedByGuard(state, isLittleGirlProtectedByGuard) {
+            state.preferences.game.options.roles.littleGirl.isProtectedByGuard = isLittleGirlProtectedByGuard;
+        },
+        setPreferenceGameOptionCanGuardProtectTwice(state, canGuardProtectTwice) {
+            state.preferences.game.options.roles.guard.canProtectTwice = canGuardProtectTwice;
+        },
+        setPreferenceGameOptionAncientLivesCountAgainstWerewolves(state, ancientLivesCountAgainstWerewolves) {
+            state.preferences.game.options.roles.ancient.livesCountAgainstWerewolves = ancientLivesCountAgainstWerewolves;
+        },
+        setPreferenceGameOptionDoesAncientTakeHisRevenge(state, doesAncientTakeHisRevenge) {
+            state.preferences.game.options.roles.ancient.doesTakeHisRevenge = doesAncientTakeHisRevenge;
+        },
+        setPreferenceGameOptionDoesIdiotDieOnAncientDeath(state, doesIdiotDieOnAncientDeath) {
+            state.preferences.game.options.roles.idiot.doesDieOnAncientDeath = doesIdiotDieOnAncientDeath;
+        },
+        setPreferenceGameOptionSistersWakingUpInterval(state, sistersWakingUpInterval) {
+            state.preferences.game.options.roles.twoSisters.wakingUpInterval = sistersWakingUpInterval;
+        },
+        setPreferenceGameOptionBrothersWakingUpInterval(state, brothersWakingUpInterval) {
+            state.preferences.game.options.roles.threeBrothers.wakingUpInterval = brothersWakingUpInterval;
+        },
+        setPreferenceGameOptionIsFoxPowerlessIfMissesWerewolf(state, isFoxPowerlessIfMissesWerewolf) {
+            state.preferences.game.options.roles.fox.isPowerlessIfMissesWerewolf = isFoxPowerlessIfMissesWerewolf;
+        },
+        setPreferenceGameOptionDoesBearTamerGrowlIfInfected(state, doesBearTamerGrowlIfInfected) {
+            state.preferences.game.options.roles.bearTamer.doesGrowlIfInfected = doesBearTamerGrowlIfInfected;
+        },
+        setPreferenceGameOptionIsWildChildTransformationRevealed(state, isWildChildTransformationRevealed) {
+            state.preferences.game.options.roles.wildChild.isTransformationRevealed = isWildChildTransformationRevealed;
+        },
+        setPreferenceGameOptionIsDogWolfChosenSideRevealed(state, isDogWolfChosenSideRevealed) {
+            state.preferences.game.options.roles.dogWolf.isChosenSideRevealed = isDogWolfChosenSideRevealed;
+        },
+        setPreferenceGameOptionStutteringJudgeVoteRequestsCount(state, stutteringJudgeVoteRequestsCount) {
+            state.preferences.game.options.roles.stutteringJudge.voteRequestsCount = stutteringJudgeVoteRequestsCount;
+        },
+        setPreferenceGameOptionThiefAdditionalCardsCount(state, thiefAdditionalCardsCount) {
+            state.preferences.game.options.roles.thief.additionalCardsCount = thiefAdditionalCardsCount;
+        },
+        setPreferenceGameOptionMustThiefChooseBetweenWerewolves(state, mustThiefChooseBetweenWerewolves) {
+            state.preferences.game.options.roles.thief.mustChooseBetweenWerewolves = mustThiefChooseBetweenWerewolves;
+        },
+        setPreferenceGameOptionPiedPiperCharmedPeopleCountPerNight(state, piedPiperCharmedPeopleCountPerNight) {
+            state.preferences.game.options.roles.piedPiper.charmedPeopleCountPerNight = piedPiperCharmedPeopleCountPerNight;
+        },
+        setPreferenceGameOptionIsPiedPiperPowerlessIfInfected(state, isPiedPiperPowerlessIfInfected) {
+            state.preferences.game.options.roles.piedPiper.isPowerlessIfInfected = isPiedPiperPowerlessIfInfected;
+        },
+        setPreferenceGameOptionRavenMarkPenalty(state, markPenalty) {
+            state.preferences.game.options.roles.raven.markPenalty = markPenalty;
+        },
         saveUserPreferences(state) {
             state.preferences.save();
         },
@@ -135,6 +219,118 @@ export default {
         },
         setPreferenceSoundEffectsVolume({ commit }, volume) {
             commit("setPreferenceSoundEffectsVolume", volume);
+            commit("saveUserPreferences");
+        },
+        setPreferencesGameOptions({ commit }, gameOptions) {
+            commit("setPreferencesGameOptions", gameOptions);
+            commit("saveUserPreferences");
+        },
+        setPreferenceGameOptionRepartitionIsHidden({ commit }, isHidden) {
+            commit("setPreferenceGameOptionRepartitionIsHidden", isHidden);
+            commit("saveUserPreferences");
+        },
+        setPreferenceGameOptionAreRolesRevealedOnDeath({ commit }, areRolesRevealedOnDeath) {
+            commit("setPreferenceGameOptionAreRolesRevealedOnDeath", areRolesRevealedOnDeath);
+            commit("saveUserPreferences");
+        },
+        setPreferenceGameOptionIsSheriffEnabled({ commit }, isSheriffEnabled) {
+            commit("setPreferenceGameOptionIsSheriffEnabled", isSheriffEnabled);
+            commit("saveUserPreferences");
+        },
+        setPreferenceGameOptionSheriffElectedAtTurn({ commit }, sheriffElectedAtTurn) {
+            commit("setPreferenceGameOptionSheriffElectedAtTurn", sheriffElectedAtTurn);
+            commit("saveUserPreferences");
+        },
+        setPreferenceGameOptionSheriffElectedAtPhase({ commit }, sheriffElectedAtPhase) {
+            commit("setPreferenceGameOptionSheriffElectedAtPhase", sheriffElectedAtPhase);
+            commit("saveUserPreferences");
+        },
+        setPreferenceGameOptionIsSheriffVoteDoubled({ commit }, isSheriffVoteDoubled) {
+            commit("setPreferenceGameOptionIsSheriffVoteDoubled", isSheriffVoteDoubled);
+            commit("saveUserPreferences");
+        },
+        setPreferenceGameOptionIsBigBadWolfPowerlessIfWerewolfDies({ commit }, isBigBadWolfPowerlessIfWerewolfDies) {
+            commit("setPreferenceGameOptionIsBigBadWolfPowerlessIfWerewolfDies", isBigBadWolfPowerlessIfWerewolfDies);
+            commit("saveUserPreferences");
+        },
+        setPreferenceGameOptionWhiteWerewolfWakingUpInterval({ commit }, whiteWerewolfWakingUpInterval) {
+            commit("setPreferenceGameOptionWhiteWerewolfWakingUpInterval", whiteWerewolfWakingUpInterval);
+            commit("saveUserPreferences");
+        },
+        setPreferenceGameOptionIsSeerTalkative({ commit }, isSeerTalkative) {
+            commit("setPreferenceGameOptionIsSeerTalkative", isSeerTalkative);
+            commit("saveUserPreferences");
+        },
+        setPreferenceGameOptionCanSeerSeeRoles({ commit }, canSeerSeeRoles) {
+            commit("setPreferenceGameOptionCanSeerSeeRoles", canSeerSeeRoles);
+            commit("saveUserPreferences");
+        },
+        setPreferenceGameOptionIsLittleGirlProtectedByGuard({ commit }, isLittleGirlProtectedByGuard) {
+            commit("setPreferenceGameOptionIsLittleGirlProtectedByGuard", isLittleGirlProtectedByGuard);
+            commit("saveUserPreferences");
+        },
+        setPreferenceGameOptionCanGuardProtectTwice({ commit }, canGuardProtectTwice) {
+            commit("setPreferenceGameOptionCanGuardProtectTwice", canGuardProtectTwice);
+            commit("saveUserPreferences");
+        },
+        setPreferenceGameOptionAncientLivesCountAgainstWerewolves({ commit }, ancientLivesCountAgainstWerewolves) {
+            commit("setPreferenceGameOptionAncientLivesCountAgainstWerewolves", ancientLivesCountAgainstWerewolves);
+            commit("saveUserPreferences");
+        },
+        setPreferenceGameOptionDoesAncientTakeHisRevenge({ commit }, doesAncientTakeHisRevenge) {
+            commit("setPreferenceGameOptionDoesAncientTakeHisRevenge", doesAncientTakeHisRevenge);
+            commit("saveUserPreferences");
+        },
+        setPreferenceGameOptionDoesIdiotDieOnAncientDeath({ commit }, doesIdiotDieOnAncientDeath) {
+            commit("setPreferenceGameOptionDoesIdiotDieOnAncientDeath", doesIdiotDieOnAncientDeath);
+            commit("saveUserPreferences");
+        },
+        setPreferenceGameOptionSistersWakingUpInterval({ commit }, sistersWakingUpInterval) {
+            commit("setPreferenceGameOptionSistersWakingUpInterval", sistersWakingUpInterval);
+            commit("saveUserPreferences");
+        },
+        setPreferenceGameOptionBrothersWakingUpInterval({ commit }, brothersWakingUpInterval) {
+            commit("setPreferenceGameOptionBrothersWakingUpInterval", brothersWakingUpInterval);
+            commit("saveUserPreferences");
+        },
+        setPreferenceGameOptionIsFoxPowerlessIfMissesWerewolf({ commit }, isFoxPowerlessIfMissesWerewolf) {
+            commit("setPreferenceGameOptionIsFoxPowerlessIfMissesWerewolf", isFoxPowerlessIfMissesWerewolf);
+            commit("saveUserPreferences");
+        },
+        setPreferenceGameOptionDoesBearTamerGrowlIfInfected({ commit }, doesBearTamerGrowlIfInfected) {
+            commit("setPreferenceGameOptionDoesBearTamerGrowlIfInfected", doesBearTamerGrowlIfInfected);
+            commit("saveUserPreferences");
+        },
+        setPreferenceGameOptionIsWildChildTransformationRevealed({ commit }, isWildChildTransformationRevealed) {
+            commit("setPreferenceGameOptionIsWildChildTransformationRevealed", isWildChildTransformationRevealed);
+            commit("saveUserPreferences");
+        },
+        setPreferenceGameOptionIsDogWolfChosenSideRevealed({ commit }, isDogWolfChosenSideRevealed) {
+            commit("setPreferenceGameOptionIsDogWolfChosenSideRevealed", isDogWolfChosenSideRevealed);
+            commit("saveUserPreferences");
+        },
+        setPreferenceGameOptionStutteringJudgeVoteRequestsCount({ commit }, stutteringJudgeVoteRequestsCount) {
+            commit("setPreferenceGameOptionStutteringJudgeVoteRequestsCount", stutteringJudgeVoteRequestsCount);
+            commit("saveUserPreferences");
+        },
+        setPreferenceGameOptionThiefAdditionalCardsCount({ commit }, thiefAdditionalCardsCount) {
+            commit("setPreferenceGameOptionThiefAdditionalCardsCount", thiefAdditionalCardsCount);
+            commit("saveUserPreferences");
+        },
+        setPreferenceGameOptionMustThiefChooseBetweenWerewolves({ commit }, mustThiefChooseBetweenWerewolves) {
+            commit("setPreferenceGameOptionMustThiefChooseBetweenWerewolves", mustThiefChooseBetweenWerewolves);
+            commit("saveUserPreferences");
+        },
+        setPreferenceGameOptionPiedPiperCharmedPeopleCountPerNight({ commit }, piedPiperCharmedPeopleCountPerNight) {
+            commit("setPreferenceGameOptionPiedPiperCharmedPeopleCountPerNight", piedPiperCharmedPeopleCountPerNight);
+            commit("saveUserPreferences");
+        },
+        setPreferenceGameOptionIsPiedPiperPowerlessIfInfected({ commit }, isPiedPiperPowerlessIfInfected) {
+            commit("setPreferenceGameOptionIsPiedPiperPowerlessIfInfected", isPiedPiperPowerlessIfInfected);
+            commit("saveUserPreferences");
+        },
+        setPreferenceGameOptionRavenMarkPenalty({ commit }, markPenalty) {
+            commit("setPreferenceGameOptionRavenMarkPenalty", markPenalty);
             commit("saveUserPreferences");
         },
     },